55: extract the remaining symbols, which are related to examples, into the newly-created `frontend/src/features/examples/` directory
