+++ frontend/src/features/examples/RecordNewExample.tsx
@@ -5,14 +5,13 @@ import { ThunkDispatch } from "redux-thunk";
 import { v4 as uuidv4 } from "uuid";
 
 import { URL_FOR_FIRST_PAGE_OF_EXAMPLES } from "../../constants";
+import { IState, logOut } from "../../store";
 import {
-  IState,
-  logOut,
   ActionFetchExamples,
   fetchExamples,
   ActionCreateExample,
   createExample,
-} from "../../store";
+} from "./examplesSlice";
 
 import { IActionAlertsCreate, alertsCreate } from "../alerts/alertsSlice";
