+++ frontend/src/features/examples/Search.tsx
@@ -16,11 +16,10 @@ import {
   logOut,
   selectExamplesIds,
   selectExamplesEntities,
-  ActionFetchExamples,
-  fetchExamples,
   selectExamplesMeta,
   selectExamplesLinks,
 } from "../../store";
+import { ActionFetchExamples, fetchExamples } from "./examplesSlice";
 
 import { IActionAlertsCreate, alertsCreate } from "../alerts/alertsSlice";
